requirements.txt
httpx
rich
python-dotenv
duckduckgo-search
prompt_toolkit
